requirements.txt
blinker==1.9.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
cryptography==46.0.5
Flask==3.1.3
flask-cors==6.0.2
google-auth==2.48.0
greenlet==3.3.2
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
psycopg2-binary==2.9.10
python-dotenv==1.2.2
requests==2.32.5
rsa==4.9.1
SQLAlchemy==2.0.48
typing_extensions==4.15.0
urllib3==2.6.3
Werkzeug==3.1.6
